requirements.txt
affine==2.3.0
alabaster==0.7.12
appdirs==1.4.4
applaunchservices==0.2.1
appnope==0.1.2
argh==0.26.2
arrow==1.2.1
astroid==2.6.6
async-generator==1.10
atomicwrites==1.4.0
attrs==21.2.0
autopep8==1.6.0
Babel==2.9.1
backcall==0.2.0
backports.functools-lru-cache==1.6.4
beautifulsoup4==4.11.2
binaryornot==0.4.4
black==22.3.0
bleach==4.1.0
branca==0.4.2
brotlipy==0.7.0
certifi==2024.2.2
cffi==1.14.6
chardet==4.0.0
charset-normalizer==2.0.0
click==8.0.3
click-plugins==1.1.1
cligj==0.7.2
cloudpickle==2.0.0
CO2Signal==0.4.2
colorama==0.4.4
cookiecutter==1.7.0
cryptography==35.0.0
cycler==0.11.0
cytoolz==0.11.2
dask==2021.11.1
dataclasses==0.8
debugpy==1.5.1
decorator==5.1.0
defusedxml==0.7.1
diff-match-patch==20200713
dill==0.3.4
docutils==0.17.1
entrypoints==0.3
ephem==4.1
et-xmlfile==1.1.0
Fiona==1.8.20
flake8==3.9.2
folium==0.0.0
fsspec==2021.11.0
future==0.18.2
GDAL==3.3.3
geographiclib==1.52
geopandas==0.10.2
geopy==2.2.0
gmpy2==2.1.2
idna==3.1
imagecodecs==2021.8.26
imageio==2.9.0
imagesize==1.3.0
importlib-metadata==4.8.2
importlib-resources==5.4.0
inflection==0.5.1
intervaltree==3.0.2
ipykernel==6.5.0
ipython==7.29.0
ipython-genutils==0.2.0
isort==5.10.1
jedi==0.18.0
Jinja2==3.0.3
jinja2-time==0.2.0
joblib==1.1.0
jsonschema==4.2.1
jupyter-client==6.1.12
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
keyring==23.2.1
kiwisolver==1.3.2
knmi-py==0.1.9
lazy-object-proxy==1.6.0
line-profiler==3.3.1
llvmlite==0.37.0
locket==0.2.0
mapclassify==2.4.3
MarkupSafe==2.0.1
matplotlib==3.4.3
matplotlib-inline==0.1.3
mccabe==0.6.1
mistune==0.8.4
mpl3d==0.0.1
mpmath==1.3.0
munch==2.5.0
mypy-extensions==0.4.3
nbclient==0.5.8
nbconvert==6.3.0
nbformat==5.1.3
nest-asyncio==1.5.1
networkx==2.6.3
numba==0.54.1
numexpr==2.8.3
numpy==1.20.3
numpydoc==1.1.0
olefile==0.46
openpyxl==3.1.2
osmnx==1.1.1
OWSLib==0.25.0
packaging==21.0
pandas==1.3.4
pandocfilters==1.5.0
parso==0.8.2
partd==1.2.0
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
pip==21.3.1
platformdirs==2.3.0
pluggy==1.0.0
pooch==1.5.2
poyo==0.5.0
prompt-toolkit==3.0.22
psutil==5.8.0
ptyprocess==0.7.0
pycodestyle==2.7.0
pycparser==2.21
pydocstyle==6.1.1
pyflakes==2.3.1
Pygments==2.10.0
pylint==2.9.6
pyls-spyder==0.4.0
pyOpenSSL==21.0.0
pyparsing==3.0.6
pyproj==3.2.1
PyQt5==5.12.3
PyQt5_sip==4.19.18
PyQtChart==5.12
PyQtWebEngine==5.12.1
pyrsistent==0.18.0
PySocks==1.7.1
python-dateutil==2.8.2
python-lsp-black==1.0.0
python-lsp-jsonrpc==1.0.0
python-lsp-server==1.2.4
pytz==2021.3
PyWavelets==1.2.0
PyYAML==6.0
pyzmq==22.3.0
QDarkStyle==3.0.2
qstylizer==0.2.1
QtAwesome==1.1.0
qtconsole==5.1.1
QtPy==1.11.2
rasterio==1.2.10
rasterstats==0.16.0
regex==2021.11.10
requests==2.26.0
rope==0.21.1
Rtree==0.9.7
scikit-image==0.18.3
scikit-learn==1.0.1
scipy==1.7.2
setuptools==59.1.0
Shapely==1.8.0
simplejson==3.17.5
six==1.16.0
snowballstemmer==2.1.0
snuggs==1.4.7
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
sparse==0.13.0
Sphinx==4.3.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
spyder==5.1.5
spyder-kernels==2.1.3
sympy==1.12
tables==3.7.0
testpath==0.5.0
textdistance==4.2.2
threadpoolctl==3.0.0
three-merge==0.1.1
tifffile==2021.11.2
tinycss2==1.1.0
toml==0.10.2
tomli==1.2.2
toolz==0.11.2
tornado==6.1
tqdm==4.62.3
traitlets==5.1.1
typed-ast==1.5.0
typing-extensions==3.10.0.2
ujson==4.2.0
urllib3==1.26.7
watchdog==2.1.6
wcwidth==0.2.5
webencodings==0.5.1
wheel==0.37.0
whichcraft==0.6.1
wrapt==1.12.1
wurlitzer==3.0.2
xyzservices==2021.11.0
yapf==0.31.0
zipp==3.6.0
